+++ sub/osd_state.h
@@ -5,8 +5,6 @@
 
 #include "osd.h"
 
-#define OSD_CONV_CACHE_MAX 4
-
 enum mp_osdtype {
     OSDTYPE_SUB,
     OSDTYPE_SUB2, // IDs must be numerically successive
@@ -47,10 +45,6 @@ struct osd_object {
 
     // OSDTYPE_EXTERNAL2
     struct sub_bitmaps *external2;
-
-    // caches for OSD conversion (internal to render_object())
-    struct osd_conv_cache *cache[OSD_CONV_CACHE_MAX];
-    struct sub_bitmaps cached;
 
     // VO cache state
     int vo_change_id;
